created the SendAck() method
I added changes to processFrame() and finishFrameRecieve() that take into account the dataframe number which is now the last byte after the parity byte and before the stopTag. I did not know how to determine the frame number when creating the frame, so I just set it to zero for now (I marked it with a TODO).